Correct documentation to have the correct future/past times
I noticed that the code in the documentation was wrong.

the nbf time used a "time in the past" instead of a "time in the future" the code as written would not have raised the error being demonstrated. Same was true for the token exp demo. The exp was being set to time in the future and so the token would not yet be expired. To get the demoed error to raise, you would have to set the exp to time in the past.